// frontend/src/backend.d.ts
export interface IngredientInfo {
    suitableSkinTypes: Array<SkinType>;
    isAcneProneConcern: boolean;
    isClogProneConcern: boolean;
    safetyClassification: IngredientSafety;
    name: string;
    description: string;
    isSensitiveConcern: boolean;
}
// ...
export enum IngredientSafety {
    safe = "safe",
    conditional = "conditional",
    harmful = "harmful",
    risky = "risky"
}
export enum PriceRange {
    low = "low",
    high = "high",
    medium = "medium"
}
export enum ProductCategory {
    sunscreen = "sunscreen",
    cleanser = "cleanser",
    treatment = "treatment",
    serum = "serum",
    moisturizer = "moisturizer"
}
export enum ProductSuitability {
    caution = "caution",
    suitable = "suitable",
    not_recommended = "not_recommended"
}
export enum SkinType {
    dry = "dry",
    combination = "combination",
    normal = "normal",
    oily = "oily",
    sensitive = "sensitive"
}